Fix indexing for multiple-call extraction of accelerometer data
bmi2_extract_accel only produces correct results when all of the accelerometer samples contained in the FIFO buffer are read in one call. (I am running it in headerless mode)

Obviously, a workaround is to just call bmi2_extract_accel with a large enough buffer to fetch all the samples at once. But the existence of fields like acc_byte_start_idx in the bmi2_fifo_frame indicate to me that the library is intended to support multiple partial extraction of sensor data from the FIFO.

Equivalent changes may need to be applied to gyr, aux and for header mode.